refactor: Carve out tests
